requirements/base.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2.6
django-allauth==65.11.2
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==25.0
Pillow==11.3.0
pycparser==2.23
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
